Moved packing/unpacking functions
